added router to refactor code
